Generify scripts so that they don't need to be symlinked into the code folders

+++ build-android.sh
@@ -1,9 +1,18 @@
 #!/usr/bin/env bash
 
+if [ ! -f "client.mk" ]; then
+    echo "Error: no client.mk found; you might be running in the wrong directory!" > /dev/stderr
+    exit 1;
+fi
+
+if [ ! -d "obj-android" ]; then
+    mkdir obj-android
+fi;
+
 (make -f client.mk \
     && pushd obj-android \
     && make package \
     && popd \
     && cp obj-android/dist/fennec*.apk $HOME/zspace/builds/ \
-    && ./save-build
-) | tee build.log
+    && save-build
+) | tee obj-android/build.log
